refactor for attribute affects, all nodes affected
